Feature 및 State 컴포넌트 분리 & JSX 슬롯
