+++ 04-essentials-deep-dive/01-starting-project/src/components/CoreConcepts.jsx
@@ -1,0 +1,15 @@
+import { CORE_CONCEPTS } from '../data.js';
+import CoreConcept from './CoreConcept.jsx';
+
+export default function CoreConcepts() {
+  return (
+    <section id="core-concepts">
+      <h2>Core Concepts</h2>
+      <ul>
+        {CORE_CONCEPTS.map((conceptItem) => (
+          <CoreConcept key={conceptItem.title} {...conceptItem} />
+        ))}
+      </ul>
+    </section>
+  );
+}
