+++ 04-essentials-deep-dive/01-starting-project/src/components/Section.jsx
@@ -1,0 +1,8 @@
+export default function Section({ title, children, ...styles }) {
+    return (
+        <section {...styles}>
+            <h2>{title}</h2>
+            {children}
+        </section>
+    );
+}
