+++ 04-essentials-deep-dive/01-starting-project/src/components/TabButton.jsx
@@ -1,8 +1,11 @@
-export default function TabButton({ children, onSelect, isSelected }) {
+export default function TabButton({ children, isSelected, ...props }) {
   console.log('TABBUTTON COMPONENT EXECUTING');
   return (
     <li>
-      <button className={isSelected ? 'active' : undefined} onClick={onSelect}>
+      <button
+        className={isSelected ? 'active' : undefined}
+        {...props}
+      >
         {children}
       </button>
     </li>
