+++ 04-essentials-deep-dive/01-starting-project/src/components/Tabs.jsx
@@ -1,0 +1,8 @@
+export default function Tabs({ children, buttons }) {
+  return (
+    <>
+      <menu>{buttons}</menu>
+      {children}
+    </>
+  )
+}
